+++ LMI_OctaneShotManager_Blender/registration.py
@@ -4,7 +4,7 @@ from .icons import load_icons, unload_icons
 from .properties import OctanePointCloudProperties
 from .exporters.csv_export import LMB_OT_export_csv
 from .exporters.abc_export import LMB_OT_export_abc
-from .tags_workflow import (
+from .TAGs_Workflow.tags_workflow import (
     TagCollectionItem,
     LMB_UL_tag_collections,
     LMB_OT_tag_collection_add,
