+++ Structure
@@ -4,7 +4,8 @@ LMI_OctaneShotManager_Blender/
 ├── utils.py               # Pure Python helpers
 ├── ui.py                  # Panel layout and shared UI helpers
 ├── registration.py        # Central register() / unregister()
-├── tags_workflow.py       # Logic for tagging workflow
+├── TAGs_Workflow/
+│   └── tags_workflow.py   # Logic for tagging workflow and related helpers
 ├── icons.py               # Icon loading helpers
 ├── exporters/
 │   ├── csv_export.py      # CSV export operator & related helpers
